Renamed components, cleanup and simplification
