+++ frontend/components/lines/LinesDetailMetrics/styles.module.css
